fix(macos): return to Omi after OAuth
Open the desktop app after a successful loopback OAuth callback, with a browser fallback link, and link the onboarding GitHub copy to the official repository.

Failure-Class: none
