+++ desktop/macos/Desktop/Sources/Onboarding/SecondBrain/SBOnboardingView.swift
@@ -2,6 +2,10 @@ import AppKit
 import OmiTheme
 import SwiftUI
 
+enum SBOnboardingRepository {
+  static let url = URL(string: "https://github.com/BasedHardware/omi")!
+}
+
 /// The Second Brain conversational onboarding — a chat with Omi that streams
 /// word-by-word and performs real side-effects. Replaces the legacy wizard.
 struct SBOnboardingView: View {
@@ -175,22 +179,29 @@ struct SBOnboardingView: View {
   private var promiseWidget: some View {
     VStack(alignment: .leading, spacing: 12) {
       VStack(spacing: 0) {
-        trustRow("OPEN", "Every line of me is on GitHub.")
+        trustRow("OPEN") {
+          HStack(spacing: 0) {
+            Text("Every line of me is on ")
+            Link("GitHub", destination: SBOnboardingRepository.url)
+              .underline()
+            Text(".")
+          }
+        }
         Divider().overlay(sb.ink(.w08))
-        trustRow("PRIVATE", "Your data is encrypted, and only yours.")
+        trustRow("PRIVATE") { Text("Your data is encrypted, and only yours.") }
         Divider().overlay(sb.ink(.w08))
-        trustRow("YOURS", "Pause me from the notch. Delete anything, forever.")
+        trustRow("YOURS") { Text("Pause me from the notch. Delete anything, forever.") }
       }
       .overlay(RoundedRectangle(cornerRadius: 13).stroke(sb.ink(.w1), lineWidth: 1))
       SBInkButton(title: "Set up Omi →") { model.answerPromise() }
     }
   }
 
-  private func trustRow(_ tag: String, _ text: String) -> some View {
+  private func trustRow<Content: View>(_ tag: String, @ViewBuilder content: () -> Content) -> some View {
     HStack(alignment: .top, spacing: 12) {
       Text(tag).geistMono(size: 11.5, weight: .medium).foregroundStyle(sb.ink(.w4)).frame(
         width: 52, alignment: .leading)
-      Text(text).geist(size: 14).foregroundStyle(sb.ink(.w85))
+      content().geist(size: 14).foregroundStyle(sb.ink(.w85))
       Spacer(minLength: 0)
     }
     .padding(.horizontal, 14).padding(.vertical, 12)
