+++ desktop/macos/Desktop/Tests/OAuthLoopbackCallbackServerTests.swift
@@ -5,7 +5,10 @@ import XCTest
 
 final class OAuthLoopbackCallbackServerTests: XCTestCase {
   func testReceivesCodeAndStateFromLoopbackCallback() async throws {
-    let server = try OAuthLoopbackCallbackServer.start(expectedState: "test-state")
+    let server = try OAuthLoopbackCallbackServer.start(
+      expectedState: "test-state",
+      appOpenURL: "omi-test://open"
+    )
     defer { server.stop() }
 
     let callbackTask = Task {
@@ -15,8 +18,9 @@ final class OAuthLoopbackCallbackServerTests: XCTestCase {
     let response = try sendLoopbackRequest(port: server.port, target: "/callback?code=test-code&state=test-state")
     XCTAssertTrue(response.hasPrefix("HTTP/1.1 200 OK"))
     XCTAssertTrue(response.contains("You're signed in"))
-    XCTAssertTrue(response.contains("You can close this tab and return to Omi."))
-    XCTAssertTrue(response.contains("window.close()"))
+    XCTAssertTrue(response.contains("Opening Omi…"))
+    XCTAssertTrue(response.contains("href=\"omi-test://open\""))
+    XCTAssertTrue(response.contains("window.location.assign(openOmi.href)"))
 
     let result = try await callbackTask.value
     XCTAssertEqual(result.code, "test-code")
@@ -27,7 +31,10 @@ final class OAuthLoopbackCallbackServerTests: XCTestCase {
   }
 
   func testIgnoresInvalidAndMismatchedRequestsBeforeValidCallback() async throws {
-    let server = try OAuthLoopbackCallbackServer.start(expectedState: "expected-state")
+    let server = try OAuthLoopbackCallbackServer.start(
+      expectedState: "expected-state",
+      appOpenURL: "omi-test://open"
+    )
     defer { server.stop() }
 
     let callbackTask = Task {
@@ -52,7 +59,10 @@ final class OAuthLoopbackCallbackServerTests: XCTestCase {
   }
 
   func testProviderErrorReturnsBrandedFailurePage() async throws {
-    let server = try OAuthLoopbackCallbackServer.start(expectedState: "expected-state")
+    let server = try OAuthLoopbackCallbackServer.start(
+      expectedState: "expected-state",
+      appOpenURL: "omi-test://open"
+    )
     defer { server.stop() }
 
     let callbackTask = Task {
@@ -76,10 +86,14 @@ final class OAuthLoopbackCallbackServerTests: XCTestCase {
   }
 
   func testResponseHTMLBuilderProducesBrandedSuccessAndFailurePages() {
-    let success = OAuthLoopbackCallbackServer.responseHTML(for: .success)
+    let success = OAuthLoopbackCallbackServer.responseHTML(for: .success, appOpenURL: "omi-test://open")
     XCTAssertTrue(success.contains("<title>Signed in - Omi</title>"))
     XCTAssertTrue(success.contains("You're signed in"))
     XCTAssertTrue(success.contains("background-color: #f7f7f7"))
+    XCTAssertTrue(success.contains("Open Omi"))
+    XCTAssertTrue(success.contains("href=\"omi-test://open\""))
+    XCTAssertTrue(success.contains("window.location.assign(openOmi.href)"))
+    XCTAssertFalse(success.contains("window.close()"))
     XCTAssertFalse(success.contains("Authentication complete. You can close this tab."))
 
     let failure = OAuthLoopbackCallbackServer.responseHTML(for: .failure)
