+++ desktop/macos/Desktop/Tests/SBOnboardingRepositoryTests.swift
@@ -1,0 +1,9 @@
+import XCTest
+
+@testable import Omi_Computer
+
+final class SBOnboardingRepositoryTests: XCTestCase {
+  func testGitHubLinkUsesOfficialOmiRepository() {
+    XCTAssertEqual(SBOnboardingRepository.url.absoluteString, "https://github.com/BasedHardware/omi")
+  }
+}
