+++ desktop/macos/changelog/unreleased/20260722-oauth-return-github-link.json
@@ -1,0 +1,3 @@
+{
+  "change": "After signing in on macOS, Omi now returns to the app automatically with an Open Omi fallback; onboarding also links directly to Omi's GitHub repository."
+}
